+++ packages/plugin-hardhat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.5.0 (2023-12-04)
+
+- Add `defender.getDeployApprovalProcess` and `defender.getUpgradeApprovalProcess` functions. ([#934](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/934))
+- Deprecate `defender.getDefaultApprovalProcess` function. This function is equivalent to `defender.getUpgradeApprovalProcess`.
+
+**Note**: OpenZeppelin Defender deployments is in beta and its functionality is subject to change.
+
 ## 3.0.0-alpha.0 (2023-11-30)
 
 - Deploy proxies from OpenZeppelin Contracts 5.0.